+++ src/main/java/com/besson/arknights/block/custom/chen/LgdBookcase.java
@@ -1,0 +1,38 @@
+package com.besson.arknights.block.custom.chen;
+
+import com.besson.arknights.block.FurnitureHorizontalFacingBlock;
+import com.besson.arknights.block.ModAbstractChestBlock;
+import net.minecraft.block.Block;
+import net.minecraft.block.BlockState;
+import net.minecraft.block.ShapeContext;
+import net.minecraft.block.entity.BlockEntity;
+import net.minecraft.block.entity.BlockEntityType;
+import net.minecraft.util.math.BlockPos;
+import net.minecraft.util.shape.VoxelShape;
+import net.minecraft.world.BlockView;
+import org.jetbrains.annotations.Nullable;
+
+import java.util.function.Supplier;
+
+public class LgdBookcase extends ModAbstractChestBlock<LgdBookcaseBE> {
+
+    private static final VoxelShape SHAPE_NS = Block.createCuboidShape(-2, 0, 0, 18, 32, 16);
+    private static final VoxelShape SHAPE_WE = Block.createCuboidShape(0, 0, -2, 16, 32, 18);
+
+    public LgdBookcase(Settings settings, Supplier<BlockEntityType<? extends LgdBookcaseBE>> blockEntityTypeSupplier) {
+        super(settings, blockEntityTypeSupplier);
+    }
+
+    @Override
+    public VoxelShape getOutlineShape(BlockState state, BlockView world, BlockPos pos, ShapeContext context) {
+        return switch (state.get(FACING)){
+            case WEST,EAST -> SHAPE_WE;
+            default -> SHAPE_NS;
+        };
+    }
+
+    @Override
+    public @Nullable BlockEntity createBlockEntity(BlockPos pos, BlockState state) {
+        return new LgdBookcaseBE(pos, state);
+    }
+}