+++ src/main/java/com/besson/arknights/block/custom/chen/LgdTrashCan.java
@@ -1,0 +1,23 @@
+package com.besson.arknights.block.custom.chen;
+
+import com.besson.arknights.block.FurnitureHorizontalFacingBlock;
+import net.minecraft.block.Block;
+import net.minecraft.block.BlockState;
+import net.minecraft.block.ShapeContext;
+import net.minecraft.util.math.BlockPos;
+import net.minecraft.util.shape.VoxelShape;
+import net.minecraft.world.BlockView;
+
+public class LgdTrashCan extends FurnitureHorizontalFacingBlock{
+
+    public LgdTrashCan(Settings settings) {
+        super(settings);
+    }
+
+    private static final VoxelShape SHAPE = Block.createCuboidShape(4, 0, 4, 12, 10, 12);
+
+    @Override
+    public VoxelShape getOutlineShape(BlockState state, BlockView world, BlockPos pos, ShapeContext context) {
+        return SHAPE;
+    }
+}